requirements.txt
kafka-python
requests
python-dotenv
pytest
redis
chromadb
